refactor(sidebar): move folder workspace row builders into their own module
